Change enum to enum class

+++ src/input.cpp
@@ -49,8 +49,8 @@ void handle_input() {
 
   for(int i = 0; i < NUM_DIRECTIONS; i++) {
     if(input::is_key_pressed(input_info.direction_keys[i])) {
-      player_info.position += geom::Point (dx[player_info.direction] * player_info.speed * delta_time,
-                                           dy[player_info.direction] * player_info.speed * delta_time);
+      player_info.position += geom::Point (dx[(int)player_info.direction] * player_info.speed * delta_time,
+                                           dy[(int)player_info.direction] * player_info.speed * delta_time);
       player_info.direction = (Direction) i;
     }
   }
